Implement a workaround via text encryptor locator
The bootstrap context needs a TextEncryptor that is "locator aware"
so that it can use the cipher text to first locate an encryptor
and then use it. The EnvironmentPrefixHelper from Spring
Cloud is really useful here, but it is package private, so we
hacked this workaround into the same package. A better solution
would be to copy the code into this project.